Allow float in NumberField, refactor stories

+++ src/fields/NumberField.jsx
@@ -10,7 +10,7 @@ export default class NumberField extends React.Component {
     return (
       <InputField
         className={b.mix(className)}
-        onChange={v => onChange(isNaN(parseInt(v)) ? null : parseInt(v))}
+        onChange={v => onChange(isNaN(parseFloat(v)) ? null : parseFloat(v))}
         {...props}
       />
     )
